Glosses
The feature-based method is based on representing the semantic of a concepts as a set of mixed attributes from the semantic relations and glosses in the knowledge sources. The semantic relations include the “is-a” ‘hypernym-hypernym’ relations, holonym, inverse gloss, and meronym. Thus, the semantic of car is represented as the set of features containing the following terms: vehicle, convertible, accelerator, train, and cable car. The performance of the feature-based method depends on the factors which are related to the fidelity, continuity, and balance of the knowledge sources such as WordNet 3.1. Although there are a number of knowledge sources that have been created to contain the semantic features of the concepts in the language, they include only the classical semantic features. To designing and implementing the genetic-based approach for selecting the semantic features from the semantic knowledge. The input of this approach is the semantic features (semantic and lexical relations, and glosses) in certain lexical resources.